SLM changes to reflect a MVC design: Added directories for model, view and controller and dao under main package edu.uci.java2. Moved Staff and Defect under model, moved Login to view, DefectDAO under dao packages.
Main - modified to start MVC. Added run() and creates new
DTSMainFrame.Starting off with Staff as model, Login as a view and will
need a new controller.
			
DTSMainFrame - NEW class - Main Frame for DTS Application
			
DefectsListPanel - NEW view class - currently not implemented
but will show the defect detail info when a new defect is
requested or when an update to an existing defect is requested.
The panel exists but is is blank for now.
			
DefectDetailsPanel - NEW view class - currently just a stub. 
Will be used to display the defect details for a new or 
existing defect. 
			
DefectsDetailsMenuBtnController - NEW class -  The controller for the
Add New Menu button. When the button is pressed, the 
DefectDetailsPanel is displayed.
			
DefectListUpdateMenuBtnController - NEW class -  The controller for the
List/Update Menu button. When the button is pressed, the 
DefectsListPanel is displayed.
			
MainMenu - Mods for format sake. Added setPanel() method to be used to
change the panel that is displayed below the menu buttons (i.e. the list
or detail(new/update) panels.
			
Login - minor mod to turn on or off the use of database for now
(USE_DATABASE). I was unable to access the db so temp until I /we can
all get to the database.