DefectsList - NEW Class - model containing the list of  defects from the DB (maybe only open defects or all defects).
DefectsListRowSelectionController - NEW class - Handles mouse 
clicks of rows in the JTable in DefectsListPanel. Still not
operational.
			
DefectsListPanel - mods to move mouse listener for row selection to a
controller. 
			
MainMenu - mod to pass defectID into DisplayDefectDetailsPanel(). If the
defectID is NOT 0, this will indicate a NEW defect is to be
added (and defectID will be filled in by DB auto incr) but if non-0 the
database will be used to retrieve the defect info corresponding to the
given defectID and the panel will be filled with the defect 
data.
			
DefectsDetailsMenuBtnController - modified to add argument, defectID, to
call to DisplayDefectDetailsPanel() per change to MainMenu.