Made the DefectDetailsPanel functional.
*Found a way to parse the resolution date and restored its text field.
Also made the field uneditable. Instead, added an itemlistener to the
status combobox so that when user chooses "CLOSED", field will
automatically fill in current date. If moved back to "OPEN", field will
revert to blank.

*Disabled the submit button for "CLOSED" defects so that users can't
modify them.

*Added actionlistener to the submit button and linked it to updateDefect
in DefectDAO. The button is completely functional. Code is in the same
class because I couldn't figure out a way to do it in the controller
class. Hardest part was parsing the date from string to util.date to
sql.date so it can be updated in db. Still haven't implemented email
function. Not sure if we're going to do that in a separate email button.

*Added actionlistener to cancel button. It can now close the dialog
window.